requirements.txt
click==8.1.7
numpy==1.24.1
pytest==8.0.0
torch==2.2.0
torchaudio==2.2.0
tqdm==4.66.1
soundfile
torchmetrics
hydra-core
wandb
filelock
matplotlib
scipy
torchmetrics
librosa
pyworld
torchinfo
pysptk
fastdtw
scipy
thop
lovely-tensors
rich
ptflops
pandas
einops